baselines/webgen-agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
datasets==4.0.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
jiter==0.9.0
multidict==6.6.3
multiprocess==0.70.16
numpy==2.2.6
openai==1.78.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.2.1
propcache==0.3.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
selenium==4.32.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
